+++ internal/domain/commands/delete_cache_command.go
@@ -18,25 +18,28 @@ func NewDeleteCacheCommand(settings *entities.Settings) *DeleteCacheCommand {
 }
 
 func (it *DeleteCacheCommand) Execute(toBeDeleted []string, global bool) {
-	var foundDirectories []string
-	for _, dir := range toBeDeleted {
-		logger.Infof("Clearing all %s directories...", dir)
+	for _, pattern := range toBeDeleted {
+		logger.Infof("Clearing all %s entries...", pattern)
+		var foundPaths []string
 		_ = filepath.Walk(".", func(path string, info os.FileInfo, err error) error {
 			if err != nil {
 				return err
 			}
 
-			if info.IsDir() && strings.HasSuffix(path, dir) {
-				foundDirectories = append(foundDirectories, path)
+			if strings.HasSuffix(path, pattern) {
+				foundPaths = append(foundPaths, path)
+				if info.IsDir() {
+					return filepath.SkipDir
+				}
 			}
 			return nil
 		})
 
-		for _, dirPath := range foundDirectories {
-			logger.Infof("Removing directory: %s", dirPath)
-			err := os.RemoveAll(dirPath)
+		for _, found := range foundPaths {
+			logger.Infof("Removing: %s", found)
+			err := os.RemoveAll(found)
 			if err != nil {
-				logger.Errorf("Failed to remove directory: %s, error: %v", dirPath, err)
+				logger.Errorf("Failed to remove: %s, error: %v", found, err)
 			}
 		}
 	}